requirements-web-scraper.txt
# Additional dependencies for web scraper module

# Core web scraping (already in main requirements.txt)
# requests>=2.31.0
# beautifulsoup4>=4.12.0
# lxml>=4.9.0

# Browser automation
playwright>=1.40.0

# Async HTTP client
aiohttp>=3.9.0
aiofiles>=23.2.0

# URL parsing and robots.txt
urllib3>=2.0.0

# Configuration management
pyyaml>=6.0.1

# CLI interface
click>=8.1.7

# Logging
loguru>=0.7.2

# Testing dependencies
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
requests-mock>=1.11.0
pytest-cov>=4.1.0

# Type hints
typing-extensions>=4.8.0

# Data validation
pydantic>=2.5.0  # Optional, for advanced validation
